Rework 464xlat

+++ nat46/src/464xlatcfg.c
@@ -1,6 +1,7 @@
 /* 464xlatcfg.c
  *
  * Copyright (c) 2015 Steven Barth <cyrus@openwrt.org>
+ * Copyright (c) 2017 Hans Dedecker <dedeckeh@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2
@@ -31,28 +32,31 @@ int main(int argc, const char *argv[])
 {
 	char buf[INET6_ADDRSTRLEN], prefix[INET6_ADDRSTRLEN + 4];
 	int pid;
-	
+
 	if (argc <= 1) {
 		fprintf(stderr, "Usage: %s <name> [ifname] [ipv6prefix] [ipv4addr] [ipv6addr]\n", argv[0]);
 		return 1;
 	}
-	
+
 	snprintf(buf, sizeof(buf), "/var/run/%s.pid", argv[1]);
 	FILE *fp = fopen(buf, "r");
 	if (fp) {
-		fscanf(fp, "%d", &pid);
-		kill(pid, SIGTERM);
+		if (fscanf(fp, "%d", &pid) == 1)
+			kill(pid, SIGTERM);
+
 		unlink(buf);
 		fclose(fp);
 	}
-	
+
 	if (!argv[2])
 		return 0;
-	
+
 	if (!argv[3] || !argv[4] || !(fp = fopen(buf, "wx")))
 		return 1;
 
-	signal(SIGTERM, sighandler);
+	signal(SIGTERM, SIG_DFL);
+	setvbuf(fp, NULL, _IOLBF, 0);
+	fprintf(fp, "%d\n", getpid());
 
 	prefix[sizeof(prefix) - 1] = 0;
 	strncpy(prefix, argv[3], sizeof(prefix) - 1);
@@ -70,7 +74,7 @@ int main(int argc, const char *argv[])
 		strcat(prefix, "/96");
 		freeaddrinfo(res);
 	}
-		
+
 	int i = 0;
 	int sock;
 	struct sockaddr_in6 saddr;
@@ -98,7 +102,7 @@ int main(int argc, const char *argv[])
 		sleep(3);
 		i++;
 	} while (i < 3);
-	
+
 	struct ipv6_mreq mreq = {saddr.sin6_addr, if_nametoindex(argv[2])};
 	if (!argv[5]) {
 		if (IN6_IS_ADDR_LINKLOCAL(&mreq.ipv6mr_multiaddr))
@@ -111,21 +115,21 @@ int main(int argc, const char *argv[])
 	} else if (inet_pton(AF_INET6, argv[5], &mreq.ipv6mr_multiaddr) != 1) {
 		return 1;
 	}
-	
+
 	if (setsockopt(sock, SOL_IPV6, IPV6_JOIN_ANYCAST, &mreq, sizeof(mreq)))
 		return 3;
-	
+
 	inet_ntop(AF_INET6, &mreq.ipv6mr_multiaddr, buf, sizeof(buf));
 	fputs(buf, stdout);
 	fputc('\n', stdout);
 	fflush(stdout);
-	
+
 	FILE *nat46 = fopen("/proc/net/nat46/control", "w");
 	if (!nat46 || fprintf(nat46, "add %s\nconfig %s local.style NONE local.v4 %s/32 local.v6 %s/128 "
 			"remote.style RFC6052 remote.v6 %s\n", argv[1], argv[1], argv[4], buf, prefix) < 0 ||
 			fclose(nat46))
 		return 4;
-	
+
 	if (!(pid = fork())) {
 		fclose(fp);
 		fclose(stdin);
@@ -133,6 +137,7 @@ int main(int argc, const char *argv[])
 		fclose(stderr);
 		chdir("/");
 		setsid();
+		signal(SIGTERM, sighandler);
 		pause();
 
 		nat46 = fopen("/proc/net/nat46/control", "w");
@@ -141,8 +146,9 @@ int main(int argc, const char *argv[])
 			fclose(nat46);
 		}
 	} else {
+		rewind(fp);
 		fprintf(fp, "%d\n", pid);
 	}
-	
+
 	return 0;
 }
